Lab6: Auth+JWT/backend/requirements.txt
alembic==1.17.2
asyncpg==0.30.0
bcrypt==4.3.0
dotenv==0.9.9
fastapi==0.118.2
pillow==11.3.0
pipreqs==0.5.0
psycopg2-binary==2.9.11
pydantic_core==2.41.5
pydantic==2.12.4
PyJWT==2.10.1
SQLAlchemy==2.0.44
